requirements.txt
python-socketio[client]
asyncio
aiohttp